% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitFun.R
\name{splitr}
\alias{splitr}
\title{Split a Dataframe}
\usage{
splitr(data, label, fracs, remove, min)
}
\arguments{
\item{data}{The dataframe to be split}

\item{label}{The column name which contains the groups}

\item{fracs}{Vector of fractions to split the dataframe into, must == 1}

\item{remove}{Vector of group names to remove from dataframe}

\item{min}{Minimum number of individuals per group. Groups with < min will be removed}
}
\value{
A list of dataframes
}
\description{
This function splits a dataframe into several, smaller dataframes. The
dataframes can also be modified by removing unwanted groups, and/or setting a
minimum number of individuals per group.
}
